+++ README.md
@@ -1,11 +1,11 @@
-# SGHI miniETL Core
+# SGHI ETL Core
 
 [![Checked with pyright](https://microsoft.github.io/pyright/img/pyright_badge.svg)](https://microsoft.github.io/pyright/)
 [![Ruff](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/ruff/main/assets/badge/v2.json)](https://github.com/astral-sh/ruff)
 [![pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/pre-commit/pre-commit)
-[![License: MIT](https://img.shields.io/badge/License-MIT-blue.svg)](https://github.com/savannahghi/sghi-ml-pipeline/blob/main/LICENSE)
+[![License: MIT](https://img.shields.io/badge/License-MIT-blue.svg)](https://github.com/savannahghi/sghi-etl-core/blob/main/LICENSE)
 
-sghi-miniETL-core specifies the API of simple ETL workflows. That is, this
+`sghi-etl-core` specifies the API of simple ETL workflows. That is, this
 project defines and specifies the interfaces of the main components needed to
 implement an ETL workflow.
 
@@ -24,6 +24,6 @@ pre-commit install
 
 ## License
 
-[MIT License](https://github.com/savannahghi/sghi-miniETL-core/blob/main/LICENSE)
+[MIT License](https://github.com/savannahghi/sghi-etl-core/blob/main/LICENSE)
 
 Copyright (c) 2024, Savannah Informatics Global Health Institute
