--- test/sghi/mini_etl_tests/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the ``sghi.miniETL`` module."""
